+++ src/test/php/lang/ast/unittest/emit/PrecedenceTest.class.php
@@ -27,6 +27,20 @@ class PrecedenceTest extends EmittingTest {
         }
       }'
     );
-    $this->assertEquals('('.$t->getName().')', $t->newinstance()->run());
+    $this->assertEquals('('.$t->getName().')', $t->newInstance()->run());
+  }
+
+  #[@test]
+  public function plusplus() {
+    $t= $this->type(
+      'class <T> {
+        private $number= 1;
+
+        public function run() {
+          return ++$this->number;
+        }
+      }'
+    );
+    $this->assertEquals(2, $t->newinstance()->run());
   }
 }