+++ src/test/php/lang/ast/unittest/parse/MembersTest.class.php
@@ -1,16 +1,7 @@
 <?php namespace lang\ast\unittest\parse;
 
 use lang\ast\Type;
-use lang\ast\nodes\ClassDeclaration;
-use lang\ast\nodes\Constant;
-use lang\ast\nodes\InstanceExpression;
-use lang\ast\nodes\InvokeExpression;
-use lang\ast\nodes\Literal;
-use lang\ast\nodes\Method;
-use lang\ast\nodes\Property;
-use lang\ast\nodes\ScopeExpression;
-use lang\ast\nodes\Signature;
-use lang\ast\nodes\Variable;
+use lang\ast\nodes\{ClassDeclaration, Constant, InstanceExpression, InvokeExpression, Literal, Method, Property, ScopeExpression, Signature, Variable};
 
 class MembersTest extends ParseTest {